Fix MUIv5 issue checkboxes
