+++ wondrous-app/components/Checkbox/Checkbox.tsx
@@ -1,0 +1,25 @@
+import Image from 'next/image';
+
+import Box from '@mui/material/Box';
+import MuiCheckbox, { CheckboxProps } from '@mui/material/Checkbox';
+
+import styles from './CheckboxStyles';
+
+const CheckMarkIcon = (
+  <Box sx={styles.icon}>
+    <Image src="/images/icons/checkmark.svg" alt="checkmark" width="10px" height="10px" />
+  </Box>
+);
+
+const Checkbox = ({ checked, onChange, name, sx, ...props }: CheckboxProps) => (
+  <MuiCheckbox
+    checked={checked}
+    onChange={onChange}
+    name={name}
+    sx={{ ...sx, ...styles.checkbox }}
+    checkedIcon={CheckMarkIcon}
+    {...props}
+  />
+);
+
+export default Checkbox;
