+++ wondrous-app/components/WeekDaySelector/WeekDaySelector.tsx
@@ -5,15 +5,10 @@ import FormLabel from '@mui/material/FormLabel';
 import FormControl from '@mui/material/FormControl';
 import FormGroup from '@mui/material/FormGroup';
 import FormControlLabel from '@mui/material/FormControlLabel';
-import Checkbox from '@mui/material/Checkbox';
+
+import Checkbox from 'components/Checkbox';
 
 import styles from './WeekDaySelectorStyles';
-
-const CheckMarkIcon = (
-  <Box sx={styles.icon}>
-    <Image src="/images/icons/checkmark.svg" alt="checkmark" width="8px" height="8px" />
-  </Box>
-);
 
 const WeekDaySelector = ({ onWeekDaysChange, weekDaysSelected }) => {
   const { monday, tuesday, wednesday, thursday, friday, saturday, sunday } = weekDaysSelected;
@@ -27,99 +22,43 @@ const WeekDaySelector = ({ onWeekDaysChange, weekDaysSelected }) => {
         <FormGroup sx={styles.formGroup}>
           <FormControlLabel
             sx={styles.formControlLabel}
-            control={
-              <Checkbox
-                checked={monday}
-                onChange={onWeekDaysChange}
-                name="monday"
-                sx={styles.checkbox}
-                checkedIcon={CheckMarkIcon}
-              />
-            }
+            control={<Checkbox checked={monday} onChange={onWeekDaysChange} name="monday" />}
             label="M"
             labelPlacement="bottom"
           />
           <FormControlLabel
             sx={styles.formControlLabel}
-            control={
-              <Checkbox
-                checked={tuesday}
-                onChange={onWeekDaysChange}
-                name="tuesday"
-                sx={styles.checkbox}
-                checkedIcon={CheckMarkIcon}
-              />
-            }
+            control={<Checkbox checked={tuesday} onChange={onWeekDaysChange} name="tuesday" />}
             label="T"
             labelPlacement="bottom"
           />
           <FormControlLabel
             sx={styles.formControlLabel}
-            control={
-              <Checkbox
-                checked={wednesday}
-                onChange={onWeekDaysChange}
-                name="wednesday"
-                sx={styles.checkbox}
-                checkedIcon={CheckMarkIcon}
-              />
-            }
+            control={<Checkbox checked={wednesday} onChange={onWeekDaysChange} name="wednesday" />}
             label="W"
             labelPlacement="bottom"
           />
           <FormControlLabel
             sx={styles.formControlLabel}
-            control={
-              <Checkbox
-                checked={thursday}
-                onChange={onWeekDaysChange}
-                name="thursday"
-                sx={styles.checkbox}
-                checkedIcon={CheckMarkIcon}
-              />
-            }
+            control={<Checkbox checked={thursday} onChange={onWeekDaysChange} name="thursday" />}
             label="T"
             labelPlacement="bottom"
           />
           <FormControlLabel
             sx={styles.formControlLabel}
-            control={
-              <Checkbox
-                checked={friday}
-                onChange={onWeekDaysChange}
-                name="friday"
-                sx={styles.checkbox}
-                checkedIcon={CheckMarkIcon}
-              />
-            }
+            control={<Checkbox checked={friday} onChange={onWeekDaysChange} name="friday" />}
             label="F"
             labelPlacement="bottom"
           />
           <FormControlLabel
             sx={styles.formControlLabel}
-            control={
-              <Checkbox
-                checked={saturday}
-                onChange={onWeekDaysChange}
-                name="saturday"
-                sx={styles.checkbox}
-                checkedIcon={CheckMarkIcon}
-              />
-            }
+            control={<Checkbox checked={saturday} onChange={onWeekDaysChange} name="saturday" />}
             label="S"
             labelPlacement="bottom"
           />
           <FormControlLabel
             sx={styles.formControlLabel}
-            control={
-              <Checkbox
-                checked={sunday}
-                onChange={onWeekDaysChange}
-                name="sunday"
-                sx={styles.checkbox}
-                checkedIcon={CheckMarkIcon}
-              />
-            }
+            control={<Checkbox checked={sunday} onChange={onWeekDaysChange} name="sunday" />}
             label="S"
             labelPlacement="bottom"
           />
